requirements.txt
fal-client
Pillow
python-dotenv
requests
